Add model settings support to AI providers, introduce new DTOs, expand API documentation, and update model capabilities
- Add `modelSettings` parameter to `CreateProviderRequestDto` and `UpdateProviderBody` for atomic configuration
- Introduce `ModelSettingsDto`, `ModelSettingsItemDto`, and `PreviewProviderModelsRequestDto` for model management
- Add `AiModelCapabilities` to track vision, tool calling, and thinking support across models
- Extend `ModelDto` and `ChatSettingsDto` with `alias` and `capabilities` fields
- Add new API methods: `GetProviderModels`, `PreviewProviderModels`, and `GetXlsx`
- Introduce export DTOs: `ExportChatRequestBody`, `ExportMessageRequestBody`, `ExportChatRequestBodyFolderId`
- Add TFA-related DTOs: `TfaAppCodeDto`, `TfaConfirmDataDto`, and corresponding wrappers
- Add `XlsxReportResponseDto` and `XlsxReportResponseWrapper` for XLSX report handling
- Update webhook API documentation with additional error response codes (400, 403, 404)
- Remove obsolete fields from `EmployeeFullDto` (`Title`, `Birthday`, `Sex`, `WorkFrom`)
- Mark `ChatSettingsDto.Thinking` property as obsolete in favor of `Capabilities`
- Update solution file project GUID
- Rename API method from `GetTfaConfirmUrl` to `GetTfaConfirmData`

+++ src/DocSpace.API.SDK/Model/GetWebhookTriggers200Response.cs
@@ -46,7 +46,7 @@ namespace DocSpace.API.SDK.Model
         /// <param name="links">List of links related to the response.</param>
         /// <param name="status">HTTP status code of the response.</param>
         /// <param name="statusCode">HTTP status code of the response (duplicate of status).</param>
-        public GetWebhookTriggers200Response(Dictionary<string, int> response = default, int count = default, List<GetPortalPrices200ResponseLinksInner> links = default, int status = default, int statusCode = default)
+        public GetWebhookTriggers200Response(Dictionary<string, long> response = default, int count = default, List<GetPortalPrices200ResponseLinksInner> links = default, int status = default, int statusCode = default)
         {
             this.Response = response;
             this.Count = count;
@@ -59,7 +59,7 @@ namespace DocSpace.API.SDK.Model
         /// Gets or Sets Response
         /// </summary>
         [DataMember(Name = "response", EmitDefaultValue = false)]
-        public Dictionary<string, int> Response { get; set; }
+        public Dictionary<string, long> Response { get; set; }
 
         /// <summary>
         /// The total number of items in the response
